wasde-integration/requirements.txt
Flask==2.3.0
requests==2.31.0
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.0
gunicorn==21.2.0
